orchestrator/requirements.txt
boto3
apache-airflow[amazon]
apache-airflow-providers-amazon==8.16.0